stacks/requirements.txt
aws-cdk-lib>=2.0.0
aws-cdk.aws-codestar-alpha>=2.0.0alpha1
aws-cdk.aws-batch-alpha>=2.0.0alpha1
boto3<=1.26.55
constructs>=10.0.0
